<Fix>: Refactor code

+++ src/parkinglocation/parkingLocation.controller.ts
@@ -5,6 +5,7 @@ import { UpdateParkingLocationDto } from "./dtos/updateParkingLocation.dto";
 import { AuthGuard } from "@nestjs/passport";
 import RolesGuard from "src/role/role.guard";
 import { User } from "src/user/user.entity";
+import { SearchParkingLocationDto } from "./dtos/searchParkingLocation.dto";
 
 @Controller("parking-location")
 export class ParkingLocationController {
@@ -18,51 +19,20 @@ export class ParkingLocationController {
   ) {
     return this.parkingLocationService.create(user.id, createParkingLocationDto);
   }
-
-  @Get("/search")
-  async search(
-    @Query("lat") lat: number,
-    @Query("lng") lng: number,
-    @Query("Radius") radius: number = 10,
-    @Query("isAvailable") isAvailable: boolean = true,
-    @Query("startAt") startAt?: number,
-    @Query("endAt") endAt?: number,
-    @Query("priceStartAt") priceStartAt?: number,
-    @Query("priceEndAt") priceEndAt?: number,
-    @Query("width") width?: number,
-    @Query("height") height?: number,
-    @Query("length") length?: number,
-    @Query("type") type?: number,
-    @Query("services") services?: string
-  ) {
-    return await this.parkingLocationService.search({
-      lat,
-      lng,
-      radius,
-      isAvailable,
-      startAt,
-      endAt,
-      priceStartAt,
-      priceEndAt,
-      width,
-      height,
-      length,
-      type,
-      services,
-    });
-  }
   @Get()
-  @UseGuards(AuthGuard("jwt"), RolesGuard("admin", "partner"))
-  findAll(@Req() request: Request & { user: User }) {
+  @UseGuards(RolesGuard())
+  findAll(@Req() request: Request & { user: User }, @Query() query: SearchParkingLocationDto) {
     const { user } = request;
+    if (!user) return this.parkingLocationService.search(query);
     if (!user.partner) return this.parkingLocationService.findAll();
     return this.parkingLocationService.findAll(user.partner.id);
   }
 
   @Get(":id")
-  @UseGuards(AuthGuard("jwt"), RolesGuard("admin", "partner"))
+  @UseGuards(RolesGuard("admin", "partner"))
   findOne(@Param("id") id: string, @Req() request: Request & { user: User }) {
     const { user } = request;
+    if (!user) return this.parkingLocationService.findOne(+id);
     if (user.role.name === "admin") return this.parkingLocationService.findOne(+id);
     return this.parkingLocationService.findOne(+id, user.partner.id);
   }
@@ -75,7 +45,10 @@ export class ParkingLocationController {
   }
 
   @Delete(":id")
-  remove(@Param("id") id: string) {
-    return this.parkingLocationService.remove(+id);
+  @UseGuards(RolesGuard("admin", "partner"))
+  remove(@Param("id") id: string, @Req() request: Request & { user: User }) {
+    const { user } = request;
+    if (user.role.name === "admin") return this.parkingLocationService.remove(+id);
+    return this.parkingLocationService.remove(+id, user.partner.id);
   }
 }
